+++ source/JEngine/J_Audio.cpp
@@ -108,15 +108,15 @@ void J_Mixer::setSoundVolume (const float a_volume)
     if (soundVolume > MIX_MAX_VOLUME) { soundVolume = MIX_MAX_VOLUME; }
     else if (soundVolume < 0.0) { soundVolume = 0.0; }
 
-    Mix_Volume(-1, soundVolume);
+    Mix_Volume(-1, (int)soundVolume);
 }
 
 void J_Mixer::toggleMute ()
 {
     muted = !muted;
 
-    if (muted) { Mix_Volume(-1, 0.0); }
-    else { Mix_Volume(-1, soundVolume); }
+    if (muted) { Mix_Volume(-1, 0); }
+    else { Mix_Volume(-1, (int)soundVolume); }
 }
